Customer segmentation Project
An e-commerce company has a variety of customers. Every customer has a different set of tastes and interests and may belong to different financial levels. Therefore, it is a heavy challenge for the marketing and strategy team to decide what products it should be promoting or what kind of campaigns will lead to the most lucrative results.

Spending score is one such metric through which you can segregate customers. Spending score is not just determined by the financial situation, but it is also based on other factors such as customer behavior, the kind of products a customer buys, etc.

In this project, I basically wanted to identify the customers who will most easily converge and buy products.